Task2/arkpz-pzpi-22-6-Aranzhyi-Ruslan-task2/pyproject.toml
[tool.poetry]
name = "atark-backend"
version = "0.1.0"
description = ""
authors = ["RuslanUC <dev_ruslan_uc@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
pydantic = {extras = ["email"], version = "^2.9.2"}
uvicorn = "^0.32.0"
tortoise-orm = "^0.21.7"
aerich = "^0.7.2"
bcrypt = "^4.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.24.0"
asgi-lifespan = "^2.1.0"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
